% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_tests.R
\name{adf_tests}
\alias{adf_tests}
\title{ADF tests}
\usage{
adf_tests(
  y,
  pmax = NULL,
  lmtestorder = NULL,
  type = c("none", "constant", "trend"),
  format = "simple"
)
}
\arguments{
\item{y}{a vector or `ts` object}

\item{pmax}{the maximum lag to test; if NULL, `pmax` is automatically chosen with 12*(t/100)^(.25)}

\item{lmtestorder}{the order for the Lagrange multiplier tests for serial correlation; if NULL, `pmax` is used}

\item{type}{the model deterministics: "none" for no drift or trend (Dt=0), "constant" for drift (Dt=c), or "trend" for drift and trend (Dt=c+dt)}

\item{format}{argument to be passed to knitr::kable for displaying the results output. Possible values are `"latex"`, `"html"`, `"pipe"`, `"simple"`, `"rst"`, and `"none"`. If `"none"` is selected, no output is displayed in the console.}
}
\description{
Conducts augmented Dickey-Fuller (ADF) tests for unit roots.
}
