% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.avg.score.R
\name{bed.avg.score}
\alias{bed.avg.score}
\title{Average Score Calculation for Genomic Ranges}
\usage{
bed.avg.score(loc.gr, score.gr, score.col = "score")
}
\arguments{
\item{loc.gr}{A GRanges object representing the query genomic ranges.}

\item{score.gr}{A GRanges object representing the subject genomic ranges, which contains the scores.}

\item{score.col}{A string specifying the column name in `score.gr` from which the average value is to be calculated. The default is "score".}
}
\value{
A GRanges object (`loc.gr`) with an additional column representing the average score for each range.
}
\description{
This function computes the average score for each feature in a set of query genomic ranges (`loc.gr`) based on the scores available in a set of subject genomic ranges (`score.gr`). It uses a specified score column in `score.gr` for the calculation.
}
\examples{
# Example usage of bed.avg.score
avg_scores <- bed.avg.score(loc.gr = x, score.gr = y, score.col = "score")

}
