% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigWig_Binr.R
\name{BW_bnr}
\alias{BW_bnr}
\title{BigWig File Binning}
\usage{
BW_bnr(BW, bin = 20000)
}
\arguments{
\item{BW}{A GRanges object representing the input BigWig file.}

\item{bin}{The size of the binning window in base pairs (bp). 
Defaults to 20,000 bp (20 kb). This parameter determines the resolution of the binning.}
}
\value{
A GRanges object representing the binned data. Each range in the returned object corresponds 
        to a bin, and the associated values represent the average score of the BigWig data within that bin.
}
\description{
This function takes a BigWig file represented as a GRanges object and bins it into specified intervals. 
It is primarily used to reduce the resolution of the data for easier analysis and visualization. 
The default bin size is set to 20,000 base pairs (20 kb).
}
\examples{
# Assuming 'bigwig' is a GRanges object
# Example usage of BW_bnr with default bin size
binned_data <- BW_bnr(bigwig)
# Example with a custom bin size of 10,000 bp
binned_data_custom <- BW_bnr(bigwig, bin=1e4)

}
