% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_export.R
\name{bw_export}
\alias{bw_export}
\title{Export GRanges Metadata Columns as BigWig Files}
\usage{
bw_export(grange, prefix, vars = "all")
}
\arguments{
\item{grange}{A GRanges object whose metadata columns are to be exported.}

\item{prefix}{A string representing the prefix to be used for naming the exported BigWig files. 
This prefix is combined with the variable names to create the file names.}

\item{vars}{A character vector specifying which variables (columns) from the metadata of the GRanges 
object should be exported. If set to "all" (default), all variables are exported.}
}
\value{
This function does not return a value but writes BigWig files to the current working directory
        or specified path.
}
\description{
This function facilitates exporting variables stored in the metadata columns (mcols) of a GRanges object as BigWig files. 
Users can choose to export either all variables or a specified subset.
}
\examples{
# Assuming 'gr' is a GRanges object with metadata columns
# Export all variables in metadata columns as BigWig files with the prefix "sample"
bw_export(gr, prefix = "sample")

# Export only specific variables
bw_export(gr, prefix = "sample", vars = c("var1", "var2"))

}
\keyword{BigWig,}
\keyword{GRanges,}
\keyword{export}
\keyword{genomics,}
