% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_perc.R
\name{bw_perc}
\alias{bw_perc}
\title{Convert BigWig Values to Percentiles}
\usage{
bw_perc(bigwig, cut_off = FALSE, bigwig_export = FALSE)
}
\arguments{
\item{bigwig}{A BigWig file path or a GRanges object representing the BigWig data.}

\item{cut_off}{A numeric value representing the percentile cutoff threshold. 
Scores below this percentile will be excluded. If FALSE (default), no cutoff is applied.}

\item{bigwig_export}{A logical flag indicating whether to export the result as a BigWig file. 
If TRUE, the resulting data is exported. If FALSE (default), the function returns a GRanges object.}
}
\value{
If `bigwig_export` is FALSE, returns a GRanges object with scores transformed to percentiles.
        If `bigwig_export` is TRUE, exports a BigWig file and returns NULL.
}
\description{
This function transforms the score values in a BigWig file into their corresponding percentiles. 
Users have the option to apply a cutoff threshold and to export the resulting data back into a BigWig file.
}
\examples{
# Convert BigWig scores to percentiles without a cutoff and return as a GRanges object
bw_perc("path/to/bigwig.bw")

# Apply a 50th percentile cutoff and return as a GRanges object
bw_perc("path/to/bigwig.bw", cut_off = 0.50)

# Convert to percentiles and export as a new BigWig file
bw_perc("path/to/bigwig.bw", bigwig_export = TRUE)

}
\keyword{BigWig,}
\keyword{genomics,}
\keyword{percentile,}
\keyword{transformation}
