% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_slopr.R, R/bw_slopr_etc.R
\name{BW_slopr}
\alias{BW_slopr}
\title{Slope Calculation of TSA Signals}
\usage{
BW_slopr(BW, bin, slop_wind, abs_value = TRUE)

BW_slopr(BW, bin, slop_wind, abs_value = TRUE)
}
\arguments{
\item{BW}{input BigWig file - GRanges object}

\item{bin}{bin window. Defaults to 20000 bp}

\item{slop_wind}{The number of bins to include in the slope calculation window.}

\item{deg}{Degree of the polynomial used for slope calculation. 
Typically, deg=3 is used for cubic polynomial fitting.}
}
\value{
A GRanges object with calculated slopes for each bin.

binned GRanges
}
\description{
This function calculates the slope of TSA signals (or any other continuous genomic signals)
using a cubic polynomial. It works by moving a window along the genome, 
fitting a cubic polynomial within that window, and computing the slope at the central bin.

This Function approximates TSA signal slope (or any other continous genomic signal) by a cubic polynomial. 
The way it works is that it move a window along the genome and fit cubic polynomial in that window and reads the slope at the centeral bin.
}
