% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_slopr_lin.R
\name{BW_slopr_lin}
\alias{BW_slopr_lin}
\title{TSA slopr with linear fit}
\usage{
BW_slopr_lin(BW, bin, slop_wind)
}
\arguments{
\item{BW}{input BigWig file - GRanges object}

\item{bin}{bin window. Defaults to 20000 bp}
}
\value{
binned GRanges
}
\description{
This Function approximates TSA signal slope (or any other continous genomic signal) by a linear (degree 1) polynomial.
The way it works is that it move a window along the genome and fit cubic polynomial in that window and reads the slope at the centeral bin.
}
