% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_smthr.R, R/bw_smthr_etc.R
\name{BW_smthr}
\alias{BW_smthr}
\title{Polynomial Smoothing of TSA Signals}
\usage{
BW_smthr(BW, bin, smth_wind)

BW_smthr(BW, bin, smth_wind)
}
\arguments{
\item{BW}{input BigWig file - GRanges object}

\item{bin}{bin window. Defaults to 20000 bp}

\item{smth_wind}{number of bins to include in polynomial smoothing}

\item{deg}{Degree of the polynomial used for smoothing. 
For example, use deg=1 for linear, deg=3 for cubic smoothing.}
}
\value{
A GRanges object with smoothed scores.

smooth GRanges
}
\description{
This function smoothens TSA signals (or any other continuous genomic signals)
using polynomial approximation. It operates by moving a window along the genome,
fitting a polynomial within that window, and reading the value at the central bin.
The degree of the polynomial is user-defined.

This Function approximates TSA signal (or any other continous genomic signal) by a cubic polynomial.
The way it works is that it move a window along the genome and fit cubic polynomial in that window and reads the value at the centeral bin.
}
\keyword{TSA-signal,}
\keyword{genomics,}
\keyword{smoothing}
