% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_smthr_zoo.R
\name{BW_smthr_zoo}
\alias{BW_smthr_zoo}
\title{TSA smthr}
\usage{
BW_smthr_zoo(BW, bin, smth_wind, p_deg = 3)
}
\arguments{
\item{BW}{input BigWig file - GRanges object}

\item{bin}{bin window. Defaults to 20000 bp}

\item{smth_wind}{number of bins to include in polynomial smoothing}

\item{p_deg}{polynomial degree}
}
\value{
smooth GRanges
}
\description{
This approximates TSA signal (or any other continous genomic signal) by a cubic polynomial.
The way it works is that it move a window along the genome and fit cubic polynomial in that window and reads the value at the centeral bin.
}
