% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higlass_file_prep.R
\name{higlass_file_prep}
\alias{higlass_file_prep}
\title{HiGlass_file_prep
correct chromosome sizes of BigWig file to be used in HiGlass}
\usage{
higlass_file_prep(bw_path, genome)
}
\arguments{
\item{bw_path}{path to BigWig file}

\item{genome}{human genome assembly to use? Defaults to "hg38"}
}
\description{
This function allows you to express your love of cats.
}
\examples{
higlass_file_prep(PATH_TO_BW,"hg38")
}
\keyword{HiGlass}
