% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irfs.R
\name{irfs}
\alias{irfs}
\title{Impulse response functions}
\usage{
irfs(mod, x, y, data, h = 10, cum = TRUE)
}
\arguments{
\item{mod}{a `dynlm` object}

\item{x}{a character vector of names of the independent variables to be included}

\item{y}{a character with the name of the dependent variable}

\item{data}{the data frame or matrix used for the model, for calculating standard deviations}

\item{h}{the number of lags to calculate IRFs out to}

\item{cum}{logical: whether to calculate cumulative IRFs}
}
\description{
Calculates impulse response functions (IRFs) from autoregressive distributive lag (ADL) models and generalized error correction models (GECMs). Note: this function only works for independent variables that include all lags from 1 to q.
}
