% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrms.R
\name{lrms}
\alias{lrms}
\title{Long-run multipliers}
\usage{
lrms(
  mod,
  x,
  y,
  type = c("adl", "gecm"),
  vcov = "const",
  digits = 2,
  format = "simple"
)
}
\arguments{
\item{mod}{a `dynlm` object}

\item{x}{a character vector of names of the independent variables to be included}

\item{y}{a character with the name of the dependent variable}

\item{type}{the type of model; either `"adl"` or `"gecm"`}

\item{vcov}{a character string specifying the type of heteroscedasticity-consistent covariance matrix estimation, to be passed to `sandwich::vcovHC` via `msm::deltamethod`}

\item{digits}{integer; the number of digits to round results to}
}
\description{
Calculates long-run multipliers (LRMs) from autoregressive distributive lag (ADL) models and generalized error correction models (GECMs).
}
