% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_checks.R
\name{mod_checks}
\alias{mod_checks}
\title{Diagnostic checks for time series models}
\usage{
mod_checks(mods, y, data, orders = c(6, 12, 18, 24))
}
\arguments{
\item{y}{a character with the name of the dependent variable}

\item{data}{the data frame or matrix used in the model}

\item{orders}{an integer vector of maximal orders of serial correlation to be tested}

\item{mod}{a `dynlm` object}
}
\description{
Conducts tests for model misspecification with `lmtest::resettest`, empirical fluctuation with `strucchange::efp` and `strucchange::sctest`, stationarity with the model coefficients, serial correlation with `lmtest::bgtest`, and heteroscedasticity with `lmtest::bptest`.
}
