% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_smthr_zoo.R
\name{model_fit_stats}
\alias{model_fit_stats}
\title{Model Fit Statistics}
\usage{
model_fit_stats(model)
}
\arguments{
\item{linear.model}{A \code{lm()} model.}
}
\value{
Returns a data frame with one row and a column for each statistic
}
\description{
Returns lm model fit statistics R-squared, adjucted R-squared,
     predicted R-squared and PRESS.
     Thanks to John Mount for his 6-June-2014 blog post, R style tip: prefer functions that return data frames" for
     the idea \link{http://www.win-vector.com/blog/2014/06/r-style-tip-prefer-functions-that-return-data-frames}
}
