% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_sig.R
\name{phi_sig}
\alias{phi_sig}
\title{Significance levels for phi tests}
\usage{
phi_sig(tstat, phi = c(1, 2, 3), n)
}
\arguments{
\item{tstat}{the test statistic for the phi test}

\item{phi}{an integer specifying which phi is being tested; must be 1 for \eqn{\phi_1}, 2 for \eqn{\phi_2}, or 3 for for \eqn{\phi_3}}

\item{n}{sample size (number of time points)}
}
\description{
Helps interpret phi tests from ADF procedures for unit roots.
}
\examples{
phi_sig(tstat=3.654, phi=2, n=150)
}
