% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_tests.R
\name{phi_tests}
\alias{phi_tests}
\title{ADF tests with uncertain deterministics}
\usage{
phi_tests(
  y,
  pmax = NULL,
  lmtestorder = NULL,
  lag.strategy = c("p", "aic", "bic", "diffs"),
  format = "simple"
)
}
\arguments{
\item{y}{a vector or `ts` object}

\item{pmax}{the maximum lag to test; if NULL, `pmax` is automatically chosen with 12*(t/100)^(.25)}

\item{lmtestorder}{the order for the Lagrange multiplier tests for serial correlation; if NULL, `pmax` is used}

\item{lag.strategy}{the strategy for choosing lags `"p"` (for smallest lags with white noise residuals), `"aic"`, `"bic"`, or `"diffs"` (first significant differences)}

\item{format}{argument to be passed to knitr::kable for displaying the results output. Possible values are `"latex"`, `"html"`, `"pipe"`, `"simple"`, `"rst"`, and `"none"`. If `"none"` is selected, no output is displayed in the console.}
}
\description{
Conducts augmented Dickey-Fuller (ADF) tests for unit roots.
}
