% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_smthr_zoo.R
\name{PRESS}
\alias{PRESS}
\title{PRESS}
\usage{
PRESS(linear.model)
}
\arguments{
\item{linear.model}{A linear regression model (class 'lm'). Required.

calculate the predictive residuals
calculate the PRESS}
}
\description{
Returns the PRESS statistic (predictive residual sum of squares).
             Useful for evaluating predictive power of regression models.
}
\author{
Thomas Hopper
}
