% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psig.R
\name{psig}
\alias{psig}
\title{Asterisk generator for p-values}
\usage{
psig(p, stars = 2, dot = TRUE, blank = TRUE)
}
\arguments{
\item{p}{a numeric vector of probabilities}

\item{stars}{an integer indicating the maximum number of stars to show}

\item{dot}{logical: if true, returns "." for significance at .05 < p < 0.10}

\item{blank}{logical: if true, returns "" for insignificant values; if false, returns NA for insignificant values}
}
\description{
Returns asterisks based on level of significance
}
\examples{
psig(c(.03, .0015, .6, .0002, .09))
psig(c(.03, .0015, .6, .0002, .09), dot=FALSE)
psig(c(.03, .0015, .6, .0002, .09), blank=FALSE)
}
