% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_coefs.R
\name{rename_coefs}
\alias{rename_coefs}
\title{Rename covariates from ADL and GECM models}
\usage{
rename_coefs(mod)
}
\arguments{
\item{mod}{a `dynlm` object}
}
\description{
Makes coefficient names more readable.
}
