% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slp_center_finder.R
\name{slp_center_find}
\alias{slp_center_find}
\title{High Slope Center Finder}
\usage{
slp_center_find(
  bw_slp_path,
  quantile_cutoff = 0.9,
  method = "local_center",
  center_width = 20000,
  export = TRUE
)
}
\arguments{
\item{bw_slp_path}{Path to the input file with slope data, typically generated 
by the BW_slopr function.}

\item{quantile_cutoff}{The quantile used to determine high slope regions. 
Defaults to 0.9, meaning the top 10% of slopes are considered high.}

\item{method}{Method for identifying the center of high-slope regions. 
Options are 'local_center' (finds local maxima within high slope regions) 
and 'simple_center' (takes the central position of high slope regions). 
Default is 'local_center'.}

\item{center_width}{Width of the region around the center to be considered. 
Defaults to 20,000 bp (2e4).}

\item{export}{Boolean indicating whether to export the results. 
If TRUE, results are saved as a .bed file. Defaults to TRUE.}
}
\value{
A GRanges object representing the center of high-slope regions.
}
\description{
This function identifies the center of high-slope regions in genomic data. 
Two methods are available for finding the center: 'simple_center' and 'local_center'. 
The latter, uses the second derivative of smoothed TSA-seq data to find these regions, 
previously identified using the BW_slopr function.
}
\examples{
centers <- slp_center_find("example.bw", quantile_cutoff=0.9, method="local_center", center_width=2e4)

}
\seealso{
`BW_slopr`
}
\keyword{TSA-seq}
\keyword{genomics,}
