% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ur_tests.R
\name{ur_tests}
\alias{ur_tests}
\title{Unit root tests}
\usage{
ur_tests(
  y,
  pmax = NULL,
  lmtestorder = NULL,
  type = c("none", "constant", "trend"),
  tests = c("adf", "pp", "dfgls", "kpss"),
  format = "simple"
)
}
\arguments{
\item{y}{a vector or `ts` object}

\item{pmax}{the maximum lag to test; if NULL, `pmax` is automatically chosen with 12*(t/100)^(.25)}

\item{lmtestorder}{the order for the Lagrange multiplier tests for serial correlation; if NULL, `pmax` is used}

\item{type}{the model deterministics}

\item{tests}{a character vector specifying which tests to conduct: "adf" for augmented Dickey-Fuller tests, "pp" for Phillips-Perron tests, "dfgls" for Dickey-Fuller tests with generalized least squares, and "kpss" for Kwiatkowski-Phillips-Schmidt-Shin tests. Can include anywhere from one to four of these tests; defaults to all four.}

\item{format}{aargument to be passed to knitr::kable for displaying the results output. Possible values are `"latex"`, `"html"`, `"pipe"`, `"simple"`, `"rst"`, and `"none"`. If `"none"` is selected, no output is displayed in the console.}
}
\description{
Conducts multiple types of tests for unit roots and automatically selects the best fitting models to display.
}
