% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecm_se.R
\name{vecm_se}
\alias{vecm_se}
\title{Standard errors for VECMs}
\usage{
vecm_se(cajo, rls)
}
\arguments{
\item{cajo}{a `ca.jo` object}

\item{rls}{a `cajorls` object}
}
\description{
Calculates standard errors for coefficients from vector error correction models (VECMs) fitted with `urca::ca.jo`
}
